requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
SQLAlchemy==2.0.46
pydantic==2.12.5
psycopg2-binary==2.9.9
python-dotenv==1.1.1
